update helios to 0.10.2
