--- workers/beerus/beerus_web_bg.wasm.d.ts
@@ -1,18 +1,0 @@
-/* tslint:disable */
-/* eslint-disable */
-export const memory: WebAssembly.Memory;
-export function __wbg_beerus_free(a: number, b: number): void;
-export function beerus_new(a: number, b: number, c: number): number;
-export function beerus_get_state(a: number): number;
-export function beerus_execute(a: number, b: number, c: number): number;
-export function set_panic_hook(): void;
-export function __wbindgen_malloc(a: number, b: number): number;
-export function __wbindgen_realloc(a: number, b: number, c: number, d: number): number;
-export const __wbindgen_export_2: WebAssembly.Table;
-export const __wbindgen_export_3: WebAssembly.Table;
-export function closure1197_externref_shim(a: number, b: number, c: number): void;
-export function __wbindgen_free(a: number, b: number, c: number): void;
-export function __wbindgen_exn_store(a: number): void;
-export function __externref_table_alloc(): number;
-export function closure3242_externref_shim(a: number, b: number, c: number, d: number): void;
-export function __wbindgen_start(): void;
