--- workers/helios/helios_ts_bg.wasm.d.ts
@@ -1,123 +1,0 @@
-/* tslint:disable */
-/* eslint-disable */
-export const memory: WebAssembly.Memory;
-export const __wbg_ethereumclient_free: (a: number, b: number) => void;
-export const ethereumclient_new: (a: number, b: number, c: number, d: number, e: number, f: number, g: number, h: number, i: number, j: number, k: number, l: number) => [number, number, number];
-export const ethereumclient_sync: (a: number) => any;
-export const ethereumclient_subscribe: (a: number, b: any, c: number, d: number, e: any) => any;
-export const ethereumclient_unsubscribe: (a: number, b: number, c: number) => [number, number, number];
-export const ethereumclient_wait_synced: (a: number) => any;
-export const ethereumclient_chain_id: (a: number) => number;
-export const ethereumclient_get_block_number: (a: number) => any;
-export const ethereumclient_get_balance: (a: number, b: any, c: any) => any;
-export const ethereumclient_get_transaction_by_hash: (a: number, b: number, c: number) => any;
-export const ethereumclient_get_transaction_by_block_hash_and_index: (a: number, b: any, c: any) => any;
-export const ethereumclient_get_transaction_by_block_number_and_index: (a: number, b: any, c: any) => any;
-export const ethereumclient_get_transaction_count: (a: number, b: any, c: any) => any;
-export const ethereumclient_get_block_transaction_count_by_hash: (a: number, b: any) => any;
-export const ethereumclient_get_block_transaction_count_by_number: (a: number, b: any) => any;
-export const ethereumclient_get_block_by_number: (a: number, b: any, c: number) => any;
-export const ethereumclient_get_block_by_hash: (a: number, b: number, c: number, d: number) => any;
-export const ethereumclient_get_code: (a: number, b: any, c: any) => any;
-export const ethereumclient_get_storage_at: (a: number, b: any, c: any, d: any) => any;
-export const ethereumclient_get_proof: (a: number, b: any, c: any, d: any) => any;
-export const ethereumclient_call: (a: number, b: any, c: any) => any;
-export const ethereumclient_estimate_gas: (a: number, b: any, c: any) => any;
-export const ethereumclient_create_access_list: (a: number, b: any, c: any) => any;
-export const ethereumclient_gas_price: (a: number) => any;
-export const ethereumclient_max_priority_fee_per_gas: (a: number) => any;
-export const ethereumclient_send_raw_transaction: (a: number, b: number, c: number) => any;
-export const ethereumclient_get_transaction_receipt: (a: number, b: any) => any;
-export const ethereumclient_get_block_receipts: (a: number, b: any) => any;
-export const ethereumclient_get_logs: (a: number, b: any) => any;
-export const ethereumclient_get_filter_changes: (a: number, b: any) => any;
-export const ethereumclient_get_filter_logs: (a: number, b: any) => any;
-export const ethereumclient_uninstall_filter: (a: number, b: any) => any;
-export const ethereumclient_new_filter: (a: number, b: any) => any;
-export const ethereumclient_new_block_filter: (a: number) => any;
-export const ethereumclient_new_pending_transaction_filter: (a: number) => any;
-export const ethereumclient_client_version: (a: number) => any;
-export const __wbg_lineaclient_free: (a: number, b: number) => void;
-export const lineaclient_new: (a: number, b: number, c: number, d: number) => [number, number, number];
-export const lineaclient_sync: (a: number) => any;
-export const lineaclient_wait_synced: (a: number) => any;
-export const lineaclient_chain_id: (a: number) => number;
-export const lineaclient_get_block_number: (a: number) => any;
-export const lineaclient_get_balance: (a: number, b: any, c: any) => any;
-export const lineaclient_get_transaction_by_hash: (a: number, b: number, c: number) => any;
-export const lineaclient_get_transaction_by_block_hash_and_index: (a: number, b: any, c: any) => any;
-export const lineaclient_get_transaction_by_block_number_and_index: (a: number, b: any, c: any) => any;
-export const lineaclient_get_transaction_count: (a: number, b: any, c: any) => any;
-export const lineaclient_get_block_transaction_count_by_hash: (a: number, b: any) => any;
-export const lineaclient_get_block_transaction_count_by_number: (a: number, b: any) => any;
-export const lineaclient_get_block_by_number: (a: number, b: any, c: number) => any;
-export const lineaclient_get_block_by_hash: (a: number, b: number, c: number, d: number) => any;
-export const lineaclient_get_code: (a: number, b: any, c: any) => any;
-export const lineaclient_get_storage_at: (a: number, b: any, c: any, d: any) => any;
-export const lineaclient_get_proof: (a: number, b: any, c: any, d: any) => any;
-export const lineaclient_call: (a: number, b: any, c: any) => any;
-export const lineaclient_estimate_gas: (a: number, b: any, c: any) => any;
-export const lineaclient_create_access_list: (a: number, b: any, c: any) => any;
-export const lineaclient_gas_price: (a: number) => any;
-export const lineaclient_max_priority_fee_per_gas: (a: number) => any;
-export const lineaclient_send_raw_transaction: (a: number, b: number, c: number) => any;
-export const lineaclient_get_transaction_receipt: (a: number, b: any) => any;
-export const lineaclient_get_block_receipts: (a: number, b: any) => any;
-export const lineaclient_get_logs: (a: number, b: any) => any;
-export const lineaclient_get_filter_changes: (a: number, b: any) => any;
-export const lineaclient_get_filter_logs: (a: number, b: any) => any;
-export const lineaclient_uninstall_filter: (a: number, b: any) => any;
-export const lineaclient_new_filter: (a: number, b: any) => any;
-export const lineaclient_new_block_filter: (a: number) => any;
-export const lineaclient_new_pending_transaction_filter: (a: number) => any;
-export const lineaclient_client_version: (a: number) => any;
-export const lineaclient_subscribe: (a: number, b: any, c: number, d: number, e: any) => any;
-export const lineaclient_unsubscribe: (a: number, b: number, c: number) => [number, number, number];
-export const __wbg_opstackclient_free: (a: number, b: number) => void;
-export const opstackclient_new: (a: number, b: number, c: number, d: number, e: number, f: number) => [number, number, number];
-export const opstackclient_sync: (a: number) => any;
-export const opstackclient_wait_synced: (a: number) => any;
-export const opstackclient_chain_id: (a: number) => number;
-export const opstackclient_get_block_number: (a: number) => any;
-export const opstackclient_get_balance: (a: number, b: any, c: any) => any;
-export const opstackclient_get_transaction_by_hash: (a: number, b: number, c: number) => any;
-export const opstackclient_get_transaction_by_block_hash_and_index: (a: number, b: any, c: any) => any;
-export const opstackclient_get_transaction_by_block_number_and_index: (a: number, b: any, c: any) => any;
-export const opstackclient_get_transaction_count: (a: number, b: any, c: any) => any;
-export const opstackclient_get_block_transaction_count_by_hash: (a: number, b: any) => any;
-export const opstackclient_get_block_transaction_count_by_number: (a: number, b: any) => any;
-export const opstackclient_get_block_by_number: (a: number, b: any, c: number) => any;
-export const opstackclient_get_block_by_hash: (a: number, b: number, c: number, d: number) => any;
-export const opstackclient_get_code: (a: number, b: any, c: any) => any;
-export const opstackclient_get_storage_at: (a: number, b: any, c: any, d: any) => any;
-export const opstackclient_get_proof: (a: number, b: any, c: any, d: any) => any;
-export const opstackclient_call: (a: number, b: any, c: any) => any;
-export const opstackclient_estimate_gas: (a: number, b: any, c: any) => any;
-export const opstackclient_create_access_list: (a: number, b: any, c: any) => any;
-export const opstackclient_gas_price: (a: number) => any;
-export const opstackclient_max_priority_fee_per_gas: (a: number) => any;
-export const opstackclient_send_raw_transaction: (a: number, b: number, c: number) => any;
-export const opstackclient_get_transaction_receipt: (a: number, b: any) => any;
-export const opstackclient_get_block_receipts: (a: number, b: any) => any;
-export const opstackclient_get_logs: (a: number, b: any) => any;
-export const opstackclient_get_filter_changes: (a: number, b: any) => any;
-export const opstackclient_get_filter_logs: (a: number, b: any) => any;
-export const opstackclient_uninstall_filter: (a: number, b: any) => any;
-export const opstackclient_new_filter: (a: number, b: any) => any;
-export const opstackclient_new_block_filter: (a: number) => any;
-export const opstackclient_new_pending_transaction_filter: (a: number) => any;
-export const opstackclient_client_version: (a: number) => any;
-export const opstackclient_subscribe: (a: number, b: any, c: number, d: number, e: any) => any;
-export const opstackclient_unsubscribe: (a: number, b: number, c: number) => [number, number, number];
-export const __wbindgen_malloc: (a: number, b: number) => number;
-export const __wbindgen_realloc: (a: number, b: number, c: number, d: number) => number;
-export const __wbindgen_exn_store: (a: number) => void;
-export const __externref_table_alloc: () => number;
-export const __wbindgen_export_4: WebAssembly.Table;
-export const __wbindgen_free: (a: number, b: number, c: number) => void;
-export const __wbindgen_export_6: WebAssembly.Table;
-export const __externref_table_dealloc: (a: number) => void;
-export const closure279_externref_shim: (a: number, b: number, c: any) => void;
-export const _dyn_core__ops__function__FnMut_____Output___R_as_wasm_bindgen__closure__WasmClosure___describe__invoke__h1fdd452e26e5f037: (a: number, b: number) => void;
-export const closure224_externref_shim: (a: number, b: number, c: any, d: any) => void;
-export const __wbindgen_start: () => void;
